extracted attributes
