+++ OpenTelemetry.sln
@@ -309,7 +309,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Options", "Options", "{4949
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shims", "Shims", "{A0CB9A10-F22D-4E66-A449-74B3D0361A9C}"
 	ProjectSection(SolutionItems) = preProject
+		CallerArgumentExpressionAttribute.cs = CallerArgumentExpressionAttribute.cs
 		src\Shared\Shims\IsExternalInit.cs = src\Shared\Shims\IsExternalInit.cs
+		NotNullAttribute.cs = NotNullAttribute.cs
 		src\Shared\Shims\NullableAttributes.cs = src\Shared\Shims\NullableAttributes.cs
 	EndProjectSection
 EndProject
